pyproject.toml
[project]
name = "duckduckgo-mcp"
version = "0.1.1"
description = "MCP Server for searching via DuckDuckGo"
readme = "README.md"
authors = [{ name = "Luca Gioacchini", email = "lucagioack@gmail.com" }]
requires-python = ">=3.10"
dependencies = [
    "beautifulsoup4>=4.13.3", 
    "httpx>=0.28.1", 
    "mcp[cli]>=1.3.0",
    "openai>=1.0.0",
    "python-dotenv==1.0.1",
    "fastapi==0.123.10"
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]